source code handling code rng fixes opencl rng code vim youcompleteme configuration

+++ rng.h
@@ -1,8 +1,10 @@
 #ifndef RNG_H
 #define RNG_H
 
-void rngSeed(unsigned long seed);
-unsigned int rngGenerate32();
+#include <stdint.h>
+
+void rngSeed(uint32_t seed);
+uint32_t rngGenerate32();
 
 static inline float rngGenerateFloat(float min, float max) {
 	float diff = max - min;
